Live Update requests/friends
