+++ src/lib/socketManager.ts
@@ -1,8 +1,11 @@
-import { socket as importedSocket, friendNowPlaying } from '$lib/store';
+import { socket as importedSocket, friendNowPlaying, friends, pendingRequests, SavedUser } from '$lib/store';
 import { get } from 'svelte/store';
 import { toast } from 'svelte-sonner';
 import { browser } from '$app/environment';
 import { UserManager } from './api/UserManager';
+import type { AddedFriend } from './types/addedfriends';
+
+let fetchedFriends: [] = [];
 
 export const socketManager = () => {
 	if (browser) {
@@ -21,6 +24,7 @@ export const socketManager = () => {
 					},
 				}
 			);
+			pendingRequests.set(data);
 		});
 
 		socket?.on('notFound', async (data) => {
@@ -36,6 +40,8 @@ export const socketManager = () => {
 			const id = data.id;
 			const friend = await UserManager.getUserbyId(id);
 			toast.success(friend.name + ' (' + friend.username + ') accepted your friend request!');
+			fetchedFriends = await UserManager.getFriends();
+			sortFriends(fetchedFriends);
 		});
 
 		socket?.on('nowPlaying', async (data) => {
@@ -46,3 +52,23 @@ export const socketManager = () => {
 		});
 	}
 };
+
+
+async function sortFriends(unsorted: any) {
+	let newFriends = unsorted.map((friend: { user_id: any; friend_id: any; }) => {
+		if (friend.user_id === get(SavedUser).id) {
+			return { user_id: get(SavedUser).id, friend_id: friend.friend_id };
+		} else {
+			return { user_id: get(SavedUser).id, friend_id: friend.user_id };
+		}
+	});
+	newFriends.forEach(async (friend: any) => {
+		const friendData = await UserManager.getUserbyId(friend.friend_id);
+		const newFriend: AddedFriend = {
+			id: friendData.id,
+			name: friendData.name,
+			username: friendData.username
+		}
+		friends.set([...get(friends), newFriend]);
+	});
+}
